requirements.txt
absl-py==2.0.0
affine==2.4.0
ants==0.0.7
antspyx==0.5.3
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blobfile==3.0.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.4
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.16
efficientnet_pytorch==0.7.1
einops==0.8.0
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.12.4
fonttools==4.55.0
fqdn==1.5.1
fsspec==2023.10.0
gdown==4.7.1
gevent==24.11.1
greenlet==3.1.1
h5py==3.12.1
huggingface-hub==0.27.1
idna==3.4
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.4.0
jupyter-console==6.6.3
jupyter_core==5.4.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.9
lpips==0.1.4
lxml==5.3.0
MarkupSafe==2.1.3
matplotlib==3.9.1
matplotlib-inline==0.1.6
mistune==3.0.2
ml_collections==0.1.1
monai==1.4.0
mpi4py==3.1.4
mpmath==1.3.0
munch==4.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2
nibabel==5.3.2
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
overrides==7.4.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
patsy==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pip==24.2
platformdirs==3.11.0
pretrainedmodels==0.7.4
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodomex==3.21.0
Pygments==2.16.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
rasterio==1.4.3
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
safetensors==0.5.2
scikit-image==0.24.0
scipy==1.11.3
segmentation_models_pytorch==0.4.0
Send2Trash==1.8.2
setuptools==75.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.5.2
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.12
terminado==0.17.1
tifffile==2024.8.30
timm==1.0.14
tinycss2==1.2.1
tomli==2.0.1
torch==2.1.0
torchmetrics==1.4.0.post0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2024.2
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
wheel==0.44.0
widgetsnbextension==4.0.9
zipp==3.21.0
zope.event==5.0
zope.interface==7.1.1
